[chore] refactor range parts calculation
