+++ internal/controllers/helpers/handlers.go
@@ -4,11 +4,13 @@ import (
 	"encoding/json"
 	"strconv"
 
+	"github.com/DIMO-Network/device-definitions-api/pkg/grpc"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/gofiber/fiber/v2"
 	"github.com/golang-jwt/jwt/v4"
 	"github.com/pkg/errors"
 	"github.com/rs/zerolog"
+	"github.com/volatiletech/null/v8"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 	//d "github.com/dexidp/dex/api/v2"
@@ -148,4 +150,51 @@ func ErrorHandler(c *fiber.Ctx, err error, logger zerolog.Logger, isProduction b
 		"code":    code,
 		"message": err.Error(),
 	})
+}
+
+type DeviceAttributeType string
+
+const (
+	MPG                 DeviceAttributeType = "mpg"
+	FuelTankCapacityGal DeviceAttributeType = "fuel_tank_capacity_gal"
+	MpgHighway          DeviceAttributeType = "mpg_highway"
+)
+
+func GetActualDeviceDefinitionMetadataValues(dd *grpc.GetDeviceDefinitionItemResponse, deviceStyleID null.String) (float64, float64, float64) {
+
+	var fuelTankCapGal, mpg, mpgHwy float64 = 0, 0, 0
+
+	var metadata []*grpc.DeviceTypeAttribute
+
+	if !deviceStyleID.IsZero() {
+		for _, style := range dd.DeviceStyles {
+			if style.Id == deviceStyleID.String {
+				metadata = style.DeviceAttributes
+				break
+			}
+		}
+	}
+
+	if len(metadata) == 0 && dd != nil && dd.DeviceAttributes != nil {
+		metadata = dd.DeviceAttributes
+	}
+
+	for _, attr := range metadata {
+		switch DeviceAttributeType(attr.Name) {
+		case FuelTankCapacityGal:
+			if v, err := strconv.ParseFloat(attr.Value, 32); err == nil {
+				fuelTankCapGal = v
+			}
+		case MPG:
+			if v, err := strconv.ParseFloat(attr.Value, 32); err == nil {
+				mpg = v
+			}
+		case MpgHighway:
+			if v, err := strconv.ParseFloat(attr.Value, 32); err == nil {
+				mpgHwy = v
+			}
+		}
+	}
+
+	return fuelTankCapGal, mpg, mpgHwy
 }
