+++ internal/controllers/user_devices_controller.go
@@ -1350,24 +1350,10 @@ func (udc *UserDevicesController) GetRange(c *fiber.Ctx) error {
 		RangeSets: []RangeSet{},
 	}
 	udd := userDevice.R.UserDeviceData
-	if len(dds) > 0 && dds[0].DeviceAttributes != nil && len(udd) > 0 {
-		var fuelTankCapGal, mpg, mpgHwy float64
-		for _, attr := range dds[0].DeviceAttributes {
-			switch attr.Name {
-			case "fuel_tank_capacity_gal":
-				if v, err := strconv.ParseFloat(attr.Value, 32); err == nil {
-					fuelTankCapGal = v
-				}
-			case "mpg":
-				if v, err := strconv.ParseFloat(attr.Value, 32); err == nil {
-					mpg = v
-				}
-			case "mpg_highway":
-				if v, err := strconv.ParseFloat(attr.Value, 32); err == nil {
-					mpgHwy = v
-				}
-			}
-		}
+	if len(dds) > 0 && dds[0] != nil && len(udd) > 0 {
+
+		fuelTankCapGal, mpg, mpgHwy := helpers.GetActualDeviceDefinitionMetadataValues(dds[0], userDevice.DeviceStyleID)
+
 		sortByJSONFieldMostRecent(udd, "fuelPercentRemaining")
 		fuelPercentRemaining := gjson.GetBytes(udd[0].Data.JSON, "fuelPercentRemaining")
 		dataUpdatedOn := gjson.GetBytes(udd[0].Data.JSON, "timestamp").Time()
